+++ skyplane/experimental/api/client.py
@@ -4,12 +4,12 @@ from pathlib import Path
 from typing import Optional
 
 from skyplane import tmp_log_dir
-from skyplane.api.auth_config import AWSConfig, AzureConfig, GCPConfig
-from skyplane.api.transfer_config import TransferConfig
-from skyplane.api.dataplane import Dataplane
-from skyplane.api.impl.path import parse_path
-from skyplane.api.impl.planner import DirectPlanner
-from skyplane.api.impl.provisioner import Provisioner
+from skyplane.experimental.api.auth_config import AWSConfig, AzureConfig, GCPConfig
+from skyplane.experimental.api.transfer_config import TransferConfig
+from skyplane.experimental.api.dataplane import Dataplane
+from skyplane.experimental.api.impl.path import parse_path
+from skyplane.experimental.api.impl.planner import DirectPlanner
+from skyplane.experimental.api.impl.provisioner import Provisioner
 from skyplane.obj_store.object_store_interface import ObjectStoreInterface
 from skyplane.utils import logger
