+++ tests/unit_nocloud/test_common.py
@@ -1,4 +1,4 @@
-from skyplane.api.impl.path import parse_path
+from skyplane.experimental.api.impl.path import parse_path
 
 
 def test_parse_path():
